feat: prepare for fly.io deployment - fix database structure, update dependencies, configure fly.toml
