+++ backend/app/core/config.py
@@ -16,6 +16,7 @@ class Settings(BaseSettings):
         # Эта опция позволяет Pydantic не падать, если .env файл не найден
         # (что актуально для production, где мы используем переменные окружения)
         env_file_encoding = 'utf-8'
+        extra = 'ignore'  # Игнорировать дополнительные поля
 
 # Оборачиваем создание объекта Settings в функцию с кэшированием.
 # Это значит, что Settings() будет создан только ОДИН раз при первом вызове,
