+++ backend/app/models/__init__.py
@@ -1,0 +1,5 @@
+from .user import User
+from .deck import Deck
+from .card import Card
+
+__all__ = ["User", "Deck", "Card"]
