+++ backend/app/models/deck.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, Integer, String, ForeignKey
-from app.db import Base
+from app.database import Base
 
 class Deck(Base):
     __tablename__ = "decks"
